requirements.txt
absl-py==0.6.1
astor==0.7.1
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
Cython==0.29
decorator==4.3.2
Django==2.1.2
djangorestframework==3.9.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
idna==2.7
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
nbformat==4.4.0
numexpr==2.6.8
numpy==1.15.3
plotly==3.6.1
protobuf==3.6.1
psycopg2-binary==2.7.5
pydot==1.2.4
pyparsing==2.3.0
python-dateutil==2.8.0
pytz==2018.7
PyYAML==3.13
requests==2.20.0
retrying==1.3.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tables==3.4.4
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.24
Werkzeug==0.14.1
whitenoise==4.1
